fix(aigateway): guard OpenAI cache runtime mutations
Refs: OME-884
